+++ src/main/java/esprit/tn/kaddem/repositories/UniversiteRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UniversiteRepository extends JpaRepository<Universite,Long> {
+public interface UniversiteRepository extends JpaRepository<Universite,Integer> {
 
 }
