+++ src/main/java/esprit/tn/kaddem/services/IEquipeService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface IEquipeService {
     public Equipe addEquipe(Equipe c);
     public Equipe updateEquipe(Equipe c);
-    public Equipe getById(long IdEquipe);
+    public Equipe getById(Integer IdEquipe);
     public List<Equipe> getAllEquipe();
-    public void removeEquipe(long IdEquipe);
+    public void removeEquipe(Integer IdEquipe);
 }
